+++ src/main/java/insane96mcp/iguanatweaksreborn/event/LivingHurtPreAbsorptionEvent.java
@@ -1,0 +1,22 @@
+package insane96mcp.iguanatweaksreborn.event;
+
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraftforge.event.entity.living.LivingEvent;
+
+public class LivingHurtPreAbsorptionEvent extends LivingEvent {
+    private final DamageSource source;
+    private float amount;
+    public LivingHurtPreAbsorptionEvent(LivingEntity entity, DamageSource source, float amount)
+    {
+        super(entity);
+        this.source = source;
+        this.amount = amount;
+    }
+
+    public DamageSource getSource() { return source; }
+
+    public float getAmount() { return amount; }
+
+    public void setAmount(float amount) { this.amount = amount; }
+}