+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PlayerMixin.java
@@ -1,10 +1,10 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
-import insane96mcp.iguanatweaksreborn.module.combat.RegeneratingAbsorption;
+import insane96mcp.iguanatweaksreborn.event.SREventFactory;
+import insane96mcp.iguanatweaksreborn.module.combat.stats.Stats;
 import insane96mcp.iguanatweaksreborn.module.experience.PlayerExperience;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import net.minecraft.core.BlockPos;
-import net.minecraft.tags.DamageTypeTags;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
@@ -14,9 +14,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import javax.annotation.Nullable;
@@ -56,11 +54,26 @@ public abstract class PlayerMixin extends LivingEntity {
 		return efficiency + EnchantmentsFeature.getEfficiencyBonus(efficiency, lvl);
 	}
 
-	@ModifyVariable(method = "actuallyHurt", at = @At(value = "STORE", ordinal = 0), ordinal = 1)
+	/*@ModifyVariable(method = "actuallyHurt", at = @At(value = "STORE", ordinal = 0), ordinal = 1)
 	private float onCalculateAbsorption(float f1, DamageSource damageSource, float amount) {
 		if (RegeneratingAbsorption.damageTypeTagOnly() && (damageSource.getEntity() == null || damageSource.is(DamageTypeTags.BYPASSES_ARMOR))) {
 			return amount;
 		}
 		return Math.max(amount - this.getAbsorptionAmount(), 0.0F);
+	}*/
+
+	@ModifyVariable(method = "actuallyHurt", at = @At(value = "STORE", ordinal = 2), argsOnly = true, ordinal = 0)
+	public float onPreAbsorptionCalculation(float amount, DamageSource damageSource) {
+		return SREventFactory.onLivingHurtPreAbsorption(this, damageSource, amount);
+	}
+
+	@ModifyConstant(method = "attack", constant = @Constant(floatValue = 0.2f, ordinal = 0))
+	public float attackStrengthAtMaxCooldown(float value) {
+        return Stats.noDamageWhenSpamming() ? 0f : value;
+    }
+
+	@ModifyConstant(method = "attack", constant = @Constant(floatValue = 0.8f, ordinal = 0))
+	public float attackStrengthAtFullSwing(float value) {
+		return Stats.noDamageWhenSpamming() ? 1f : value;
 	}
 }